+++ playground/lib/src/main-named.js
@@ -1,0 +1,4 @@
+export const foo = 'foo'
+
+// Force esbuild spread helpers
+console.log({ ...foo })